requirements.txt
Django==3.2.10
djangorestframework==3.12.4
dj-rest-auth==2.1.11
Pillow==8.4.0
python-dotenv==0.19.2
psycopg2-binary==2.9.3
gunicorn==20.1.0